Built-in Distributions

Built In Distributions

Omega comes with a number of built-in probability distributions.

Univariate Distributions

Omega.bernoulli — Function.

Bernoulli with weight p

source

Omega.boolbernoulli — Function.

Bool - valued Bernoulli distribution (as opposed to Float64 valued)

source

Omega.betarv — Function.

Beta distribution (alias β) parameters α and β

source

Omega.categorical — Function.

Categorical distribution with probability weight vector p

source

Omega.constant — Function.

Constant Random Variable

source

Omega.exponential — Function.

Exponential Distribution with λ

source

Omega.gammarv — Function.

Gamma distribution (alias Γ)

source

Omega.inversegamma — Function.

Inverse Gamma distribution

source

Omega.kumaraswamy — Function.

Kumaraswamy distribution, similar to beta but easier

source

Omega.logistic — Function.

Logistic Distribution

source

Omega.poisson — Function.

Poisson distribution with rate parameter λ

source

Omega.normal — Function.

Normal Distribution with mean μ and variance σ

source

Omega.uniform — Function.

Uniform distribution with lower bound a and upper bound b

source

Uniform sample from vector

source

Discrete uniform distribution with range range

source

Omega.rademacher — Function.

Rademacher distribution

source

Multivariate Distributions

Omega.mvnormal — Function.

Multivariate Normal Distribution with mean vector μ and covariance Σ

source

Omega.dirichlet — Function.

Dirichlet distribution

source

Describe distributional functions

Omega comes with some functions which summarize an entire distribution. Most of these are inherited from Distributions.jl

mean
prob