Built-in Distributions

Built In Distributions

Omega comes with a number of built-in probability distributions.

Univariate Distributions

Omega.Prim.bernoulli — Function.

bernoulli(p)
bernoulli(p, T)

Bernoulli distribution with weight p and return type T

source

Missing docstring.

Missing docstring for betarv. Check Documenter's build log for details.

Missing docstring.

Missing docstring for categorical. Check Documenter's build log for details.

Omega.constant — Function.

constant(c) Constant random variable which always outputs c

source

Missing docstring.

Missing docstring for exponential. Check Documenter's build log for details.

Missing docstring.

Missing docstring for gammarv. Check Documenter's build log for details.

Missing docstring.

Missing docstring for invgammarv. Check Documenter's build log for details.

Missing docstring.

Missing docstring for kumaraswamy. Check Documenter's build log for details.

Missing docstring.

Missing docstring for logistic. Check Documenter's build log for details.

Missing docstring.

Missing docstring for poisson. Check Documenter's build log for details.

Missing docstring.

Missing docstring for normal. Check Documenter's build log for details.

Omega.Prim.uniform — Function.

Discrete uniform distribution over unit range range

source

Discrete uniform distribution over array

source

Missing docstring.

Missing docstring for rademacher. Check Documenter's build log for details.

Multivariate Distributions

Missing docstring.

Missing docstring for mvnormal. Check Documenter's build log for details.

Missing docstring.

Missing docstring for dirichlet. Check Documenter's build log for details.

Describe distributional functions

Omega comes with some functions which summarize an entire distribution. Most of these are inherited from Distributions.jl

Missing docstring.

Missing docstring for mean. Check Documenter's build log for details.

Omega.Prim.prob — Function.

Probability x is true

source