Conditional Inference

Inference

The primary purpose of building a probabilistic model is to use it for inference.

Conditional Sampling Algorithms

While there are several kinds of things you would like to know about a conditional distribution (e.g., its mode) currently, all inference algorithms perform conditional sampling only. To sample from a conditional distribution: pass two random variables to rand, the distribution you want to sample from, and the predicate you want to condition on. For example:

weight = β(2.0, 2.0)
x = bernoulli()
rand(weight, x == 0)

It is fine to condition random variables on equalities or inequalities:

x1 = normal(0.0, 1.0)
x2 = normal(0.0, 10)
rand((x1, x2), x1 + x2 > == 0.0)

Note: to sample from more than one random variable, just pass a tuple of RandVars to rand, e.g.:

Base.rand — Function.

Random ω ∈ Ω

source

Sample n from x

source

Sample 1 from x

source

n samples from x with rejection sampling

source

Sample ω | y == true with Metropolis Hasting

source

Sample from x | y == true with Single Site Metropolis Hasting

source

Sample from x | y == true with Hamiltonian Monte Carlo

source

Sample from x | y == true with Hamiltonian Monte Carlo

source

rand(s::Sampleable)

Generate one sample for s.

rand(s::Sampleable, n::Int)

Generate n samples from s. The form of the returned object depends on the variate form of s:

When s is univariate, it returns a vector of length n.

When s is multivariate, it returns a matrix with n columns.

When s is matrix-variate, it returns an array, where each element is a sample matrix.

rand(d::UnivariateDistribution)

Generate a scalar sample from d. The general fallback is quantile(d, rand()).

rand(d::UnivariateDistribution, n::Int) -> Vector

Generates a vector of n random scalar samples from d. The general fallback is to pick random samples from sampler(d).

rand(d::MultivariateDistribution)

Sample a vector from the distribution d.

rand(d::MultivariateDistribution, n::Int) -> Vector

Sample n vectors from the distribution d. This returns a matrix of size (dim(d), n), where each column is a sample.

rand(rng::AbstractRNG, d::AbstractMvNormal)
rand(rng::AbstractRNG, d::AbstractMvNormal, n::Int)
rand!(rng::AbstractRNG, d::AbstractMvNormal, x::AbstractArray)

Sample from distribution d using the random number generator rng.

rand(d::MatrixDistribution, n)

Draw a sample matrix from the distribution d.

rand(d::Union{UnivariateMixture, MultivariateDistribution})

Draw a sample from the mixture model d.

rand(d::Union{UnivariateMixture, MultivariateMixture}, n)

Draw n samples from d.

Conditioning with cond

rand(x, y) is simply a shorter way of saying rand(cond(x, y)). That is, the primary mechanism for inference in Omega is conditioning random variables using cond.

Omega.cond — Function.

Condition random variable x with random predicate RandVar which returns Bool or SoftBool.

x = normal(0.0, 1.0)
x_ = cond(x, x > 0)

source

Condition random variable with predicate: cond(x, p) = cond(x, p(x)) cond(poisson(0.5), iseven

source

Condition within a function

function x_(ω)
  x = 0.0
  xs = Float64[]
  while bernoulli(ω, 0.8, Bool)
    x += uniform(ω, -5.0, 5.0)
    cond(ω, x <=ₛ 1.0)
    cond(ω, x >=ₛ -1.0)
    push!(xs, x)
  end
  xs
end

x = ciid(x_)
samples = rand(x, 100; alg = SSMH)

source

Conditioning the Prior

In Bayesian inference the term posterior distribution is often used instead of conditional distribution. Mathematically they are the same objec, but posterior alludes to the fact that it is the distribution after (post) observing data. Prior to observing data, your distribution is the prior.

However, conditioning is a more general concept than observing data, and we can meaningfully "condition the prior".

For example we can make a truncated normal distribution with:

x = normal(0.0, 1.0)
x_ = cond(x, x > 0.0)

A shorter way to write this is to pass a unary function as the second argument to cond

x = cond(normal(0.0, 1.0), rv -> rv > 0.0)

Or suppose we want a poisson distribution over the even numbers

cond(poission(1.0), iseven)

Conditions Propagate

When you compose condiitoned random variables together, their conditions propagate. That is, if x is conditioned, and y is conditioned, and z depends on x and y, then z inherits all the conditions of x and y automatically. For example:

ispos(x) = x > 0.0
weight = cond(normal(72, 1.0), ispos)
height = cond(normal(1.78, 1.0), ispos)
bmi = weight / height

bmi is a function of both weight and height, both of which have their own conditions (namely, they are positive). Omega automatically propagates the conditions from weight and height onto bmi respects these conditions.