Conditional Inference

Inference

Omega have several inference algorithms built in, and provides the mechanism to build your own.

Conditional Samples

If you have a random variable x and a Boolean-valued random variable y, to sample from a conditional distribution use rand(x,y).

Base.rand — Function.

Random ω ∈ Ω

source

Sample n from x

source

Sample 1 from x

source

n samples from x with rejection sampling

source

Sample ω | y == true with Metropolis Hasting

source

Sample from x | y == true with Single Site Metropolis Hasting

source

Sample from x | y == true with Hamiltonian Monte Carlo

source

Sample from x | y == true with Hamiltonian Monte Carlo

source

Sample from ω | y == true with Stochastic Gradient Hamiltonian Monte Carlo

source

Sample from x | y == true with Metropolis Hasting

source

For example:

weight = β(2.0, 2.0)
x = bernoulli()
rand(weight, x == 0)

To sample from more than one random variable, just pass a tuple of RandVars to rand, e.g.:

weight = β(2.0, 2.0)
x = bernoulli()
rand(weight, x == 0)

Conditioning with cond

rand(x, y) is simply a shorter way of saying rand(cond(x, y)). That is, the mechanism for inference in Omega is conditioning random variables:

Omega.cond — Function.

Condition random variable x with random predicate RandVar{Bool}

x = normal(0.0, 1.0)
x_ = cond(x, x > 0)

source

Condition random variable with predicate: cond(x, p) = cond(x, p(x)) cond(poisson(0.5), iseven

source

Conditioning as Prior

Conditioning Random Variables allows you to add constraints to your mode.

For example we can make a truncated normal distribution with:

x = normal(0.0, 1.0)
x_ = cond(x, x > 0.0)

A shorter way to write this is to pass a unary function as the second argument to cond

x = cond(normal(0.0, 1.0), rv -> rv > 0.0)

Or suppose we want a poisson distribution over the even numbers

cond(poission(1.0), iseven

ispos(x) = x > 0.0
weight = cond(normal(72, 1.0), ispos)
height = cond(normal(1.78, 1.0), ispos)
bmi = weight / height

bmi is a function of both weight and height, both of which have their own conditions. Omega automatically propagates the conditions from weight and height onto bmi, so that if we sample from all of them with

rand((bmi, weight, height), alg = Rejection))